+++ CHANGELOG.md
@@ -3,13 +3,18 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+- Mil Symbol sensemaker additionally checks for attributeIri in determining starting Mil symbol code
 
 ### Added
+- Added two helper methods to the Resolution sensemaker and created the AttributeCombinations and Criteria helper classes for resolution sensemaker
 - RDF endpoint displays relationships to objects
 
 ### Changed
 
 ### Fixed
+- Fixed the Resolution sensemaker's inability to process multiple attributes used for identifying duplicates
+- Handling connection errors with RabbitMQ
+- Made Mil Symbol logic user uppercase values to make config lookups case insensitive
 
 ## [0.6.1] - 2025-05-21